Benchkiller-21-05-20/Benchkiller/requirements.txt
# Core
coreapi==2.3.3
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
django-rest-swagger==2.2.0
django-split-settings==0.3.0
djangorestframework==3.9.0
Jinja2
Markdown==3.0.1
petl
# psycopg2-binary==2.7.5

# XLSX
openpyxl

# Testing
pytest
pytest-cov
hypothesis
factory_boy

# Code Style
pylint
black

# Git
pre-commit
